Moved GeneratePolygons to its own folder. Still working on the MUP create/edit windows. Validation for MUP.Edit is working, but the red error box isn't displaying on load.
