+++ VSProjects/Geomapmaker/Config.daml
@@ -42,7 +42,6 @@
           <button refID="Geomapmaker_ContactsFaults" />
         </group>
         <group id="Geomapmaker_MUPs" caption="Map Unit Polys" keytip="MUP">
-          <button refID="Geomapmaker_MapUnitPolys" />
           <button refID="Geomapmaker_MapUnitPolysCreate_Btn" />
           <button refID="Geomapmaker_MapUnitPolysEdit_Btn" />
           <button refID="Geomapmaker_GeneratePolygons_Btn" />
@@ -95,11 +94,6 @@
             Tooltip text<disabledText />
           </tooltip>
         </tool>
-        <button id="Geomapmaker_MapUnitPolys" caption="Map Unit Polys" className="Geomapmaker.ViewModels.MapUnitPolys.ShowMapUnitPolys" loadOnClick="true" smallImage="pack://application:,,,/ArcGIS.Desktop.Resources;component/Images/GenericButtonRed16.png" largeImage="pack://application:,,,/ArcGIS.Desktop.Resources;component/Images/GenericButtonRed32.png" keytip="todo" condition="dataSourceSelectedCondition">
-          <tooltip heading="Tooltip Heading">
-            Tooltip text<disabledText />
-          </tooltip>
-        </button>
         <tool id="Geomapmaker_SelectContactsFaultsTool" caption="SelectContactsFaultsTool" className="Geomapmaker.MapTools.SelectContactsFaultsTool" loadOnClick="true" smallImage="pack://application:,,,/ArcGIS.Desktop.Resources;component/Images/GenericButtonBlue16.png" largeImage="pack://application:,,,/ArcGIS.Desktop.Resources;component/Images/GenericButtonBlue32.png" keytip="todo" condition="esri_mapping_mapPane">
           <tooltip heading="Tooltip Heading">
             Tooltip text<disabledText />
@@ -130,7 +124,7 @@
             Tooltip text<disabledText />
           </tooltip>
         </button>
-        <button id="Geomapmaker_GeneratePolygons_Btn" caption="Generate All Polygons" className="Geomapmaker.ViewModels.MapUnitPolys.ShowGeneratePolygons" loadOnClick="true" smallImage="pack://application:,,,/ArcGIS.Desktop.Resources;component/Images/GenericButtonBlue16.png" largeImage="pack://application:,,,/ArcGIS.Desktop.Resources;component/Images/GenericButtonBlue32.png" keytip="todo" condition="dataSourceSelectedCondition">
+        <button id="Geomapmaker_GeneratePolygons_Btn" caption="Generate All Polygons" className="Geomapmaker.ViewModels.GeneratePolygons.ShowGeneratePolygons" loadOnClick="true" smallImage="pack://application:,,,/ArcGIS.Desktop.Resources;component/Images/GenericButtonBlue16.png" largeImage="pack://application:,,,/ArcGIS.Desktop.Resources;component/Images/GenericButtonBlue32.png" keytip="todo" condition="dataSourceSelectedCondition">
           <tooltip heading="Tooltip Heading">
             Tooltip text<disabledText />
           </tooltip>
