+++ VSProjects/Geomapmaker/Views/GeneratePolygons/GeneratePolygons.xaml
@@ -1,4 +1,4 @@
-﻿<controls:ProWindow x:Class="Geomapmaker.Views.MapUnitPolys.GeneratePolygons"
+﻿<controls:ProWindow x:Class="Geomapmaker.Views.GeneratePolygons.GeneratePolygons"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:controls="clr-namespace:ArcGIS.Desktop.Framework.Controls;assembly=ArcGIS.Desktop.Framework"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
